+++ pkg/cmd/disable/cmd.go
@@ -1,9 +1,8 @@
 // Copyright Contributors to the Open Cluster Management project
-package enable
+package disable
 
 import (
-	"github.com/stolostron/cm-cli/pkg/cmd/enable/addons"
-	"github.com/stolostron/cm-cli/pkg/cmd/enable/components"
+	"github.com/stolostron/cm-cli/pkg/cmd/disable/components"
 	genericclioptionscm "github.com/stolostron/cm-cli/pkg/genericclioptions"
 
 	"github.com/spf13/cobra"
@@ -13,11 +12,10 @@ import (
 // NewCmd provides a cobra command wrapping NewCmdImportCluster
 func NewCmd(cmFlags *genericclioptionscm.CMFlags, streams genericclioptions.IOStreams) *cobra.Command {
 	cmd := &cobra.Command{
-		Use:   "enable",
-		Short: "enable a feature",
+		Use:   "disable",
+		Short: "disable a feature",
 	}
 
-	cmd.AddCommand(addons.NewCmd(cmFlags, streams))
 	cmd.AddCommand(components.NewCmd(cmFlags, streams))
 
 	return cmd