+++ pkg/cmd/disable/components/cmd.go
@@ -1,0 +1,39 @@
+// Copyright Contributors to the Open Cluster Management project
+package components
+
+import (
+	"fmt"
+
+	genericclioptionscm "github.com/stolostron/cm-cli/pkg/genericclioptions"
+	cmdutil "k8s.io/kubectl/pkg/cmd/util"
+
+	"github.com/spf13/cobra"
+	"github.com/stolostron/cm-cli/pkg/helpers"
+	"k8s.io/cli-runtime/pkg/genericclioptions"
+)
+
+const (
+	example = `
+	# disable component
+	%[1]s disable component <component_name>
+`
+)
+
+// NewCmd ...
+func NewCmd(cmFlags *genericclioptionscm.CMFlags, streams genericclioptions.IOStreams) *cobra.Command {
+
+	o := newOptions(cmFlags, streams)
+	cmd := &cobra.Command{
+		Use:                   "component",
+		DisableFlagsInUseLine: true,
+		Short:                 "Enable components",
+		Example:               fmt.Sprintf(example, helpers.GetExampleHeader()),
+		Run: func(cmd *cobra.Command, args []string) {
+			cmdutil.CheckErr(o.complete(cmd, args))
+			cmdutil.CheckErr(o.validate())
+			cmdutil.CheckErr(o.run(streams))
+		},
+	}
+
+	return cmd
+}