+++ pkg/cmd/disable/components/options.go
@@ -1,0 +1,19 @@
+// Copyright Contributors to the Open Cluster Management project
+package components
+
+import (
+	genericclioptionscm "github.com/stolostron/cm-cli/pkg/genericclioptions"
+	"k8s.io/cli-runtime/pkg/genericclioptions"
+)
+
+type Options struct {
+	//CMFlags: The generic optiosn from the cm cli-runtime.
+	CMFlags   *genericclioptionscm.CMFlags
+	component string
+}
+
+func newOptions(cmFlags *genericclioptionscm.CMFlags, streams genericclioptions.IOStreams) *Options {
+	return &Options{
+		CMFlags: cmFlags,
+	}
+}